pyproject.toml
[project]
name = "tapo-c210-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.1.0",
    "pure-python-adb>=0.3.0.dev0",
    "pytapo>=3.3.54",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
]
